EXPERIMENTAL: FROMLIST: sched/topology: Reference the Energy Model of CPUs when available
The existing scheduling domain hierarchy is defined to map to the cache
topology of the system. However, Energy Aware Scheduling (EAS) requires
more knowledge about the platform, and specifically needs to know about
the span of Performance Domains (PD), which do not always align with
caches.

To address this issue, use the Energy Model (EM) of the system to extend
the scheduler topology code with a representation of the PDs, alongside
the scheduling domains. More specifically, a linked list of PDs is
attached to each root domain. When multiple root domains are in use,
each list contains only the PDs covering the CPUs of its root domain. If
a PD spans over CPUs of two different root domains, it will be
duplicated in both lists.

The lists are fully maintained by the scheduler from
partition_sched_domains() in order to cope with hotplug and cpuset
changes. As for scheduling domains, the list are protected by RCU to
ensure safe concurrent updates.

The linked lists should be used only from code paths protected by the
ENERGY_AWARE sched_feat.

Change-Id: I65ba204388315cd3e7048241a7dd1c119021d63e
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
(Applied from
https://lore.kernel.org/lkml/20180912091309.7551-7-quentin.perret@arm.com/)
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Udit Karode <udit.karode@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: mawrick26 <mawrick26@gmail.com>