ARM: dts: sdm845: Disable LPM wakeup prediction for the big cluster
CAF made it so that only the little cluster had prediction enabled in
commit c6157e21a5d8 ("ARM: dts: msm: Enable prediction for Silver CPUs
for SDM845"), but then later changed the way the parameters are parsed
from DT in commit fbcacdf2d53 ("cpuidle: lpm-levels: Add support to
parse LPM parameters"), rendering the property useless. They proceeded
to remove the old property to enable prediction on the little cluster
from sdm845's DT and never added the property to disable prediction on
the big cluster. However, they did do this for sdm670.

Add the missing property to disable LPM prediction on the big cluster
for power-saving and align it with sdm670.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>