ion: Use vmalloc for large scatterlist arrays
Instead of using lib/scatterlist.c's strategy of rolling an MMU with
software, it's faster to actually use our MMU and the kernel functions
designed specifically for memory allocations of this magnitude. When
allocation requests are smaller than a page, vmalloc falls back to
kmalloc; when they're larger than a page, vmalloc mucks with the page
tables in order to produce a virtually-contiguous chain of pages
satisfying the allocation. This is far more efficient than
not-invented-here MMU-like shenanigans, and improves ION's performance
considerably.

Change-Id: I4b8279d9cd7e891d3a44c088d381b34f37e5b6f7
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[0ctobot: Adapted for msm-4.9]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>